Changed structure of React Router
