+++ frontend/declaration.d.ts
@@ -1,0 +1,2 @@
+// declaration.d.ts
+declare module '*.scss';
