+++ frontend/package.json
@@ -13,7 +13,7 @@
     "keycloak-js": "^15.0.2",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
-    "react-router-dom": "6",
+    "react-router-dom": "^6.2.1",
     "react-scripts": "4.0.3",
     "web-vitals": "^1.0.1"
   },
