+++ frontend/src/App.tsx
@@ -1,7 +1,7 @@
 import "./styles/App.scss";
 import React, { FC } from "react";
 import { BrowserRouter } from "react-router-dom";
-import AppRouter from "./router/AppRouter";
+import AppRouter from "./routes";
 
 const App: FC = () => {
   return (
