+++ frontend/src/__tests__/pages/PageNotFound.test.js
@@ -1,7 +1,7 @@
 import React from "react";
 import { mountWithTheme } from "../../setupTests";
 import { MemoryRouter } from "react-router-dom";
-import AppRouter from "../../router/AppRouter";
+import AppRouter from "../../routes";
 import PageNotFound from "../../pages/PageNotFound";
 
 describe("<PageNotFound /> routing", () => {
