+++ frontend/src/components/Debug/debug.scss
