+++ frontend/src/components/DetailsList/index.tsx
@@ -1,0 +1,13 @@
+import React from "react";
+import { defaultMaxListeners } from "stream";
+
+export const DetailsList = ({ id, title }: { id?: string; title: string }) => {
+  return (
+    <>
+      <h2>
+        Details for {title}
+        {id}
+      </h2>
+    </>
+  );
+};
