+++ frontend/src/components/Footer/footer.scss
@@ -1,0 +1,6 @@
+@import "../../styles/components/variables";
+
+#BcFooter {
+  height: 75px;
+  background: $primary-color;
+}
