+++ frontend/src/components/Header/header.scss
@@ -1,0 +1,6 @@
+@import "../../styles/components/variables";
+
+#BcHeader {
+  height: 75px;
+  background: $primary-color;
+}
