--- frontend/src/components/Main/index.scss
@@ -1,14 +1,0 @@
-@import '../../styles/components/variables';
-
-.main-section {
-    width: 100%;
-    height: 100%;
-
-    header {
-        padding-top: 0.5rem;
-        padding-bottom: 0.5rem;
-        min-height: 50px;
-        background-color: $primary-color;
-        
-    }
-}
