+++ frontend/src/components/PageLink/index.js
@@ -2,10 +2,12 @@ import { Outlet, useParams } from "react-router-dom";
 import React from "react";
 
 export const PageLink = (title, outlet) => {
+
   let { id } = useParams();
   return (
     <>
       <h2>
+       
         {title} {id}
       </h2>
       {outlet && <Outlet />}
