+++ frontend/src/components/Sidebar/sideBar.scss
