--- frontend/src/index.css
@@ -1,3 +1,0 @@
-body, #root {
-    height: 100vh;
-}
