+++ frontend/src/pages/Admin/Users/users.scss
@@ -1,0 +1,86 @@
+@import "../../../styles/components/variables";
+
+.users {
+  .table-data-filter {
+    margin-top: 2.5rem;
+    margin-bottom: 1rem;
+    display: flex;
+    justify-content: space-between;
+
+    .filter-form {
+      > * {
+        margin-right: 1rem;
+      }
+
+      label {
+        font-size: 1.1rem;
+        font-weight: bold;
+        color: $nav-background;
+      }
+    }
+
+    .user-search-form {
+      input {
+        padding: 0.4rem;
+      }
+
+      button {
+        background-color: $nav-background;
+        color: $white;
+        border: none;
+        padding: 0.5rem 0.75rem;
+        font-size: 1rem;
+        border-radius: 20px;
+        font-weight: bold;
+      }
+    }
+  }
+
+  .user-table {
+    border-radius: 5px;
+    border: 2px solid $nav-background;
+
+    .user-table-head {
+      background-color: $nav-background;
+
+      .head-cell {
+        &:first-child {
+          border-right: 1px solid $white;
+        }
+
+        &:last-child {
+          min-width: 50px;
+        }
+
+        border-right: none;
+        color: $white;
+        font-weight: bold;
+        font-size: 1.2rem;
+      }
+    }
+
+    .user-table-body {
+      .table-body-row {
+        .body-cell {
+          &:first-child {
+            border-right: 2px solid $nav-background;
+          }
+
+          border-right: none;
+          border-bottom: none;
+          font-size: 1rem;
+          font-weight: bold;
+        }
+      }
+    }
+  }
+
+  .user-role-totals {
+    font-weight: bold;
+    margin-top: 1rem;
+
+    span {
+      margin-right: 1rem;
+    }
+  }
+}
