+++ frontend/src/pages/Admin/index.tsx
@@ -1,21 +1,16 @@
-import React, { FC } from "react";
-import { Outlet, useParams } from "react-router-dom";
+import React from "react";
+import { Outlet } from "react-router-dom";
 import Sidebar from "../../components/Sidebar";
 import Main from "../../components/Main";
-import Debug from "../../components/Debug";
 
-export const Admin: FC = () => {
+export const Admin = () => {
   return (
     <>
       <Sidebar />
       <Main>
-        <h2>Admin</h2>
+        <h2>Contract List</h2>
         <Outlet />
       </Main>
     </>
   );
-};
-
-export const AdminDebug: FC = () => {
-  return <Debug />;
 };
