--- frontend/src/pages/Login/index.scss
@@ -1,48 +1,0 @@
-@import '../../styles/App.scss';
-
-.login-page {
-    width: 100%;
-    height: 100%;
-    display: flex;
-    flex-direction: column;
-    justify-content: space-between;
-
-    header {
-        padding-top: 0.5rem;
-        padding-bottom: 0.5rem;
-        min-height: 50px;
-        background-color: $primary-color;
-        
-    }
-
-    main {
-        height: 300px;
-        width: 40%;
-        border-radius: 4px;
-        border: 1px solid $black;
-        align-self: center;
-        display: flex;
-        flex-direction: column;
-        justify-content: center;
-        align-items: center;
-
-        button {
-            max-width: 30%;
-            max-width: 30%;
-            padding: 0.5rem 0.75rem;
-            font-size: 1rem;
-            background-color: $primary-color;
-            border: none;
-            color: $white;
-            border-radius: 4px;
-        }
-    }
-
-    footer {
-        min-height: 50px;
-        padding-top: 0.5rem;
-        padding-bottom: 0.5rem;
-        background-color: $primary-color;
-        
-    }
-}
