--- frontend/src/pages/PageNotFound/index.tsx
@@ -1,8 +1,0 @@
-import React, { FC } from "react";
-import Main from "../../components/Main";
-
-const PageNotFound: FC = () => {
-  return <Main>Uh oh page not found.</Main>;
-};
-
-export default PageNotFound;
