+++ frontend/src/pages/Projects/index.tsx
@@ -1,19 +1,16 @@
 import React from "react";
-
-import Sidebar from "../components/Sidebar";
-import Main from "../components/Main";
 import { Outlet } from "react-router-dom";
+import Sidebar from "../../components/Sidebar";
+import Main from "../../components/Main";
 
-export const Admin = () => {
+export const Projects = () => {
   return (
     <>
       <Sidebar />
       <Main>
-        <h1>Admin</h1>
+        <h2>Project List</h2>
         <Outlet />
       </Main>
     </>
   );
 };
-
-export default Admin;
