+++ frontend/src/pages/index.js
@@ -1,0 +1,9 @@
+export { Admin } from "./Admin";
+export { Contracts } from "./Contracts";
+export { Contract } from "./Contracts/Contract";
+export { Home } from "./Home";
+export { Login } from "./Login";
+export { PageNotFound } from "./PageNotFound";
+export { Projects } from "./Projects";
+export { Project } from "./Projects/Project";
+export { Users } from "./Users";
