+++ frontend/src/routes/subRoutes/projectRoutes.tsx
@@ -2,21 +2,11 @@ import React from "react";
 import { Route } from "react-router-dom";
 import ProtectedRoute from "../ProtectedRoute";
 import { PageLink } from "../../components/PageLink";
-
-import { List } from "../../pages/Project";
+import { Project, Projects } from "../../pages";
 
 const projectRoutes = [
-  <Route key="project" path="/project" element={<ProtectedRoute component={List} />}>
-    <Route
-      path=":projectId"
-      element={
-        <ProtectedRoute
-          component={() => {
-            return PageLink("Details");
-          }}
-        />
-      }
-    />
+  <Route key="project" path="/project" element={<ProtectedRoute component={Projects} />}>
+    <Route path=":projectId" element={<ProtectedRoute component={Project} />} />
     <Route
       path=":projectId/status"
       element={
