+++ frontend/src/utils/index.tsx
@@ -1,0 +1,1 @@
+export { apiAxios } from "./apiAxios";
